locus/tracker/requirements.txt
anthropic>=0.25.0
opencv-python>=4.9.0
Pillow>=10.2.0
requests>=2.31.0
